fix: introduce lazy-loading → expense page
